+++ ui/src/app/core/accessories/accessory-info/accessory-info.component.ts
@@ -40,6 +40,7 @@ export class AccessoryInfoComponent implements OnInit {
       'Dehumidifier',
       'Television',
       'Valve',
+      'RobotVacuum',
     ],
     [
       'Door',