+++ ui/src/app/core/accessories/types/robot-vacuum/robot-vacuum.component.ts
@@ -1,0 +1,33 @@
+import { NgClass } from '@angular/common'
+import { Component, Input } from '@angular/core'
+import { TranslatePipe } from '@ngx-translate/core'
+
+import { ServiceTypeX } from '@/app/core/accessories/accessories.interfaces'
+import { LongClickDirective } from '@/app/core/directives/long-click.directive'
+
+@Component({
+  selector: 'app-robot-vacuum',
+  templateUrl: './robot-vacuum.component.html',
+  standalone: true,
+  imports: [
+    LongClickDirective,
+    NgClass,
+    TranslatePipe,
+  ],
+})
+export class RobotVacuumComponent {
+  @Input() public service: ServiceTypeX
+  @Input() public readyForControl = false
+
+  public onClick() {
+    if (!this.readyForControl) {
+      return
+    }
+
+    if ('On' in this.service.values) {
+      this.service.getCharacteristic('On').setValue(!this.service.values.On)
+    } else if ('Active' in this.service.values) {
+      this.service.getCharacteristic('Active').setValue(this.service.values.Active ? 0 : 1)
+    }
+  }
+}